+++ common/src/main/java/de/cech12/usefulhats/item/MiningHatItem.java
@@ -59,7 +59,7 @@ public class MiningHatItem extends AbstractMiningHatItem implements IEquipmentCh
 
     @Override
     public void inventoryTick(@NotNull ItemStack stack, ServerLevel level, @NotNull Entity entity, EquipmentSlot slot) {
-        if (!level.isClientSide && entity instanceof LivingEntity livingEntity) {
+        if (!level.isClientSide() && entity instanceof LivingEntity livingEntity) {
             if (!Services.REGISTRY.getEquippedHatItemStacks(livingEntity).contains(stack)) return; //only one worn stack of this item should add its effect
             //When Night Vision effect is disabled in config, do nothing.
             if (!Services.CONFIG.isMiningHatNightVisionEnabled()) return;