+++ common/src/main/java/de/cech12/usefulhats/item/PostmanHatItem.java
@@ -41,7 +41,7 @@ public class PostmanHatItem extends AbstractHatItem implements IEquipmentChangeL
 
     @Override
     public void inventoryTick(@NotNull ItemStack stack, ServerLevel level, @NotNull Entity entity, EquipmentSlot slot) {
-        if (!level.isClientSide && entity instanceof LivingEntity livingEntity) {
+        if (!level.isClientSide() && entity instanceof LivingEntity livingEntity) {
             if (!Services.REGISTRY.getEquippedHatItemStacks(livingEntity).contains(stack)) return; //only one worn stack of this item should add its effect
             int speedAmplifier = CommonLoader.getEnchantmentLevel(stack, Enchantments.EFFICIENCY);
             //When Speed effect is caused by another source, do nothing