+++ forge/src/main/java/de/cech12/usefulhats/client/UsefulHatItemExtension.java
@@ -1,7 +1,6 @@
 package de.cech12.usefulhats.client;
 
 import net.minecraft.client.model.HumanoidModel;
-import net.minecraft.client.renderer.entity.state.HumanoidRenderState;
 import net.minecraft.client.renderer.entity.state.LivingEntityRenderState;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ItemStack;
@@ -18,9 +17,9 @@ public class UsefulHatItemExtension implements IClientItemExtensions {
     @Override
     @NotNull
     public HumanoidModel<?> getHumanoidArmorModel(LivingEntityRenderState state, ItemStack itemStack, EquipmentSlot equipmentSlot, HumanoidModel<?> original) {
-        if (original instanceof HumanoidModel<?> humanoidModel) {
-            ((HumanoidModel<HumanoidRenderState>) humanoidModel).copyPropertiesTo(UsefulHatsForgeClientEvents.usefulHatModel);
-        }
+        //if (original instanceof HumanoidModel<?> humanoidModel) {
+        //    ((HumanoidModel<HumanoidRenderState>) humanoidModel).copyPropertiesTo(UsefulHatsForgeClientEvents.usefulHatModel);
+        //}
         return UsefulHatsForgeClientEvents.usefulHatModel;
     }
 }