+++ neoforge/src/main/java/de/cech12/usefulhats/client/UsefulHatItemExtension.java
@@ -2,9 +2,9 @@ package de.cech12.usefulhats.client;
 
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.Model;
-import net.minecraft.client.renderer.entity.state.HumanoidRenderState;
 import net.minecraft.client.resources.model.EquipmentClientInfo;
 import net.minecraft.world.item.ItemStack;
+import net.neoforged.neoforge.client.ClientHooks;
 import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 import org.jetbrains.annotations.NotNull;
 
@@ -19,7 +19,7 @@ public class UsefulHatItemExtension implements IClientItemExtensions {
     @NotNull
     public Model getHumanoidArmorModel(@NotNull ItemStack itemStack, @NotNull EquipmentClientInfo.LayerType layerType, @NotNull Model original) {
         if (original instanceof HumanoidModel<?> humanoidModel) {
-            ((HumanoidModel<HumanoidRenderState>) humanoidModel).copyPropertiesTo(UsefulHatsNeoForgeClientEvents.usefulHatModel);
+            ClientHooks.copyModelProperties(humanoidModel, UsefulHatsNeoForgeClientEvents.usefulHatModel);
         }
         return UsefulHatsNeoForgeClientEvents.usefulHatModel;
     }