+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [this versioning scheme](https://gist.github.com/cech12/69319028e88c50349a6b044000a6607b).
 
+## [1.21.10-6.11.0.0] - 2025-10-14
+### Changed
+- updated to Minecraft 1.21.10 (Fabric 0.135.0+1.21.10, NeoForge 21.10.16-beta, Forge 60.0.5)
+- updated Cloth Config support (20.0.148) (Fabric/Quilt)
+- updated ModMenu support (16.0.0-rc.1) (Fabric/Quilt)
+- temporary removed Accessories support until it is ported to 1.21.10 (Fabric/Quilt, NeoForge)
+
 ## [1.21.8-6.10.0.0] - 2025-10-13
 ### Changed
 - updated to Minecraft 1.21.8 (Fabric 0.134.0+1.21.8, NeoForge 21.8.47, Forge 58.1.0)